+++ contracts/1delta/modules/venus/FlashAggregator.sol
@@ -2,10 +2,11 @@
 
 pragma solidity ^0.8.28;
 
-/******************************************************************************\
-* Author: Achthar | 1delta 
-/******************************************************************************/
-
+/**
+ * \
+ * Author: Achthar | 1delta
+ * /*****************************************************************************
+ */
 import {LendingOps, IERC20} from "./VenusOps.sol";
 import {TokenTransfer} from "../../libraries/TokenTransfer.sol";
 
@@ -16,25 +17,24 @@ import {TokenTransfer} from "../../libraries/TokenTransfer.sol";
  * @notice Adds money market and default transfer functions to margin trading
  */
 contract VenusFlashAggregator is LendingOps, TokenTransfer {
-
     constructor(address _cNative, address _wNative) LendingOps(_cNative, _wNative) {}
 
-    function deposit(address underlying, uint amount) external {
+    function deposit(address underlying, uint256 amount) external {
         _transferERC20TokensFrom(underlying, msg.sender, address(this), amount);
         _deposit(underlying, amount, msg.sender);
     }
 
-    function withdraw(address underlying, uint amount) external {
+    function withdraw(address underlying, uint256 amount) external {
         _withdraw(underlying, amount, msg.sender);
         _transferERC20Tokens(underlying, msg.sender, amount);
     }
 
-    function borrow(address underlying, uint amount) external {
+    function borrow(address underlying, uint256 amount) external {
         _borrow(underlying, amount, msg.sender);
         _transferERC20Tokens(underlying, msg.sender, amount);
     }
 
-    function repay(address underlying, uint amount) external {
+    function repay(address underlying, uint256 amount) external {
         _transferERC20TokensFrom(underlying, msg.sender, address(this), amount);
         _repay(underlying, amount, msg.sender);
     }
