+++ contracts/1delta/modules/venus/IFlashAggregator.sol
@@ -3,11 +3,11 @@
 pragma solidity ^0.8.28;
 
 interface IVenusFlashAggregator {
-    function deposit(address underlying, uint amount) external;
+    function deposit(address underlying, uint256 amount) external;
 
-    function withdraw(address underlying, uint amount) external;
+    function withdraw(address underlying, uint256 amount) external;
 
-    function borrow(address underlying, uint amount) external;
+    function borrow(address underlying, uint256 amount) external;
 
-    function repay(address underlying, uint amount) external;
+    function repay(address underlying, uint256 amount) external;
 }
