+++ test/ethereum/ComposedFlashLoan.t.sol
@@ -14,7 +14,11 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
      *  borrow
      *  payback
      */
-    function test_ethereum_composed_flash_loan_open(uint8 lenderId) external /** address user, uint8 lenderId */ {
+    function test_ethereum_composed_flash_loan_open(uint8 lenderId) external 
+    /**
+     * address user, uint8 lenderId
+     */
+    {
         TestParamsOpen memory params;
         address user = testUser;
 
@@ -60,9 +64,9 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
 
         uint8 flashSource = BALANCER_V2;
         {
-            uint borrowAm = params.swapAmount +
-                (params.swapAmount * getFlashFee(flashSource)) / //
-                10000;
+            uint256 borrowAm = params.swapAmount
+                + (params.swapAmount * getFlashFee(flashSource)) //
+                    / 10000;
 
             approveBorrowDelegation(user, params.borrowAsset, borrowAm, lenderId);
 
@@ -99,7 +103,7 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
             )
         );
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
 
@@ -111,9 +115,9 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
         // deposit 10, recieve 32.1... makes 42.1...
         assertApproxEqAbs(379869471726271100564, balance, 1);
         {
-            uint borrowAm = params.swapAmount +
-                (params.swapAmount * getFlashFee(flashSource)) / //
-                10000;
+            uint256 borrowAm = params.swapAmount
+                + (params.swapAmount * getFlashFee(flashSource)) //
+                    / 10000;
             // deviations through rouding expected, accuracy for 10 decimals
             assertApproxEqAbs(borrowBalance, borrowAm, 1);
         }
@@ -149,7 +153,7 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
         uint256 borrowBalance = getBorrowBalance(user, borrowAsset, lenderId);
         uint256 balance = getCollateralBalance(user, asset, lenderId);
         bytes memory data;
-        uint witdrawAm;
+        uint256 witdrawAm;
         {
             witdrawAm = amountToFlashWithdraw + (amountToFlashWithdraw * getFlashFee(flashSource)) / 10000;
             approveWithdrawal(user, asset, witdrawAm, lenderId);
@@ -185,7 +189,7 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
             );
 
             vm.prank(user);
-            uint gas = gasleft();
+            uint256 gas = gasleft();
             IFlashAggregator(brokerProxyAddress).deltaCompose(data);
             gas = gas - gasleft();
             console.log("gas-flash-loan-close", gas);
@@ -218,21 +222,30 @@ contract ComposedFlashLoanTestEthereum is DeltaSetup {
         deal(b, address(router), 1e20);
     }
 
-    function encodeExtCall(address token, address tokenOut, address approveTarget, address target, uint amount) internal pure returns (bytes memory) {
+    function encodeExtCall(
+        address token,
+        address tokenOut,
+        address approveTarget,
+        address target,
+        uint256 amount
+    )
+        internal
+        pure
+        returns (bytes memory)
+    {
         bytes memory data = abi.encodeWithSelector(MockRouter.swapExactIn.selector, token, tokenOut, amount);
-        return
-            abi.encodePacked(
-                uint8(Commands.EXTERNAL_CALL), //
-                token,
-                approveTarget,
-                target,
-                uint112(amount),
-                uint16(data.length),
-                data
-            );
+        return abi.encodePacked(
+            uint8(Commands.EXTERNAL_CALL), //
+            token,
+            approveTarget,
+            target,
+            uint112(amount),
+            uint16(data.length),
+            data
+        );
     }
 
-    function getFlashFee(uint8 source) internal view returns (uint) {
+    function getFlashFee(uint8 source) internal view returns (uint256) {
         return source == BALANCER_V2 ? 0 : ILendingPool(AaveV3Ethereum.POOL).FLASHLOAN_PREMIUM_TOTAL();
     }
 }
